requirements.txt
flask
python-dotenv
google-genai
sumy
nltk
rouge-score
PyPDF2
